backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0
python-dotenv==1.0.0
openai==0.28.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymongo==4.6.1
